Revert "Display optional speakers of text"
