+++ assets/script.vn
@@ -1,3 +1,3 @@
 "This is an example text"
-"Narrator" "This is another example text"
-"Stranger" "So many example texts!"
+"This is another example text"
+"So many example texts!"
