+++ src/text.rs
@@ -1,13 +1,6 @@
 use bevy::prelude::*;
 
-#[derive(Component)]
-pub struct Speaker;
-
-#[derive(Component)]
-pub struct Dialog;
-
 pub struct TextEvent {
-    pub speaker: Option<String>,
     pub text: String,
 }
 
@@ -16,72 +9,34 @@ impl Plugin for TextPlugin {
     fn build(&self, app: &mut App) {
         app.add_event::<TextEvent>()
             .add_startup_system(setup)
-            .add_system(speaker_text)
-            .add_system(dialog_text);
+            .add_system(text);
     }
 }
 
 fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
-    commands
-        .spawn_bundle(
-            TextBundle::from_section(
-                "",
-                TextStyle {
-                    font: asset_server.load("fonts/dejavu/DejaVuSans.ttf"),
-                    font_size: 22.0,
-                    color: Color::WHITE,
-                },
-            )
-            .with_style(Style {
-                align_self: AlignSelf::FlexStart,
-                position_type: PositionType::Absolute,
-                position: UiRect {
-                    bottom: Val::Px(150.0),
-                    left: Val::Px(15.0),
-                    ..default()
-                },
-                ..default()
-            }),
+    commands.spawn_bundle(
+        TextBundle::from_section(
+            "",
+            TextStyle {
+                font: asset_server.load("fonts/dejavu/DejaVuSans.ttf"),
+                font_size: 22.0,
+                color: Color::WHITE,
+            },
         )
-        .insert(Speaker);
-
-    commands
-        .spawn_bundle(
-            TextBundle::from_section(
-                "",
-                TextStyle {
-                    font: asset_server.load("fonts/dejavu/DejaVuSans.ttf"),
-                    font_size: 22.0,
-                    color: Color::WHITE,
-                },
-            )
-            .with_style(Style {
-                align_self: AlignSelf::FlexStart,
-                position_type: PositionType::Absolute,
-                position: UiRect {
-                    bottom: Val::Px(120.0),
-                    left: Val::Px(15.0),
-                    ..default()
-                },
+        .with_style(Style {
+            align_self: AlignSelf::FlexStart,
+            position_type: PositionType::Absolute,
+            position: UiRect {
+                bottom: Val::Px(120.0),
+                left: Val::Px(15.0),
                 ..default()
-            }),
-        )
-        .insert(Dialog);
+            },
+            ..default()
+        }),
+    );
 }
 
-fn speaker_text(mut ev_text: EventReader<TextEvent>, mut query: Query<&mut Text, With<Speaker>>) {
-    for text_event in ev_text.iter() {
-        for mut text_display in &mut query {
-            if let Some(speaker) = text_event.speaker.clone() {
-                text_display.sections[0].value = speaker.clone();
-            } else {
-                text_display.sections[0].value = "".to_string();
-            }
-        }
-    }
-}
-
-fn dialog_text(mut ev_text: EventReader<TextEvent>, mut query: Query<&mut Text, With<Dialog>>) {
+fn text(mut ev_text: EventReader<TextEvent>, mut query: Query<&mut Text>) {
     for text_event in ev_text.iter() {
         for mut text_display in &mut query {
             text_display.sections[0].value = text_event.text.clone();
